refactor(resources/data): migrate data resources fetchers
