+++ packages/app-builder/src/components/Data/CreateTable.tsx
@@ -1,0 +1,120 @@
+import { FormErrorOrDescription } from '@app-builder/components/Form/Tanstack/FormErrorOrDescription';
+import { FormInput } from '@app-builder/components/Form/Tanstack/FormInput';
+import { FormLabel } from '@app-builder/components/Form/Tanstack/FormLabel';
+import { useLoaderRevalidator } from '@app-builder/contexts/LoaderRevalidatorContext';
+import {
+  CreateTableValue,
+  createTableValueSchema,
+  useCreateTableMutation,
+} from '@app-builder/queries/data/create-table';
+import { getFieldErrors } from '@app-builder/utils/form';
+import { useForm } from '@tanstack/react-form';
+import { useState } from 'react';
+import { useTranslation } from 'react-i18next';
+import { Button, Modal } from 'ui-design-system';
+
+export function CreateTable({ children }: { children: React.ReactNode }) {
+  const { t } = useTranslation(['data', 'navigation', 'common']);
+  const createTableMutation = useCreateTableMutation();
+  const [isOpen, setIsOpen] = useState(false);
+  const revalidate = useLoaderRevalidator();
+
+  const form = useForm({
+    defaultValues: {
+      name: '',
+      description: '',
+    } as CreateTableValue,
+    onSubmit: ({ value, formApi }) => {
+      if (formApi.state.isValid) {
+        createTableMutation.mutateAsync(value).then((result) => {
+          revalidate();
+
+          if (result.success) {
+            setIsOpen(false);
+            form.reset();
+          }
+        });
+      }
+    },
+    validators: {
+      onSubmit: createTableValueSchema,
+    },
+  });
+
+  return (
+    <Modal.Root open={isOpen} onOpenChange={setIsOpen}>
+      <Modal.Trigger asChild>{children}</Modal.Trigger>
+      <Modal.Content>
+        <form
+          onSubmit={(e) => {
+            e.preventDefault();
+            e.stopPropagation();
+            form.handleSubmit();
+          }}
+        >
+          <Modal.Title>{t('data:create_table.title')}</Modal.Title>
+          <div className="flex flex-col gap-6 p-6">
+            <div className="flex flex-1 flex-col gap-4">
+              <form.Field
+                name="name"
+                validators={{
+                  onBlur: createTableValueSchema.shape.name,
+                  onChange: createTableValueSchema.shape.name,
+                }}
+              >
+                {(field) => (
+                  <div className="flex flex-col gap-2">
+                    <FormLabel name={field.name}>{t('data:field_name')}</FormLabel>
+                    <FormInput
+                      type="text"
+                      name={field.name}
+                      defaultValue={field.state.value as string}
+                      onChange={(e) => field.handleChange(e.currentTarget.value)}
+                      onBlur={field.handleBlur}
+                      valid={field.state.meta.errors.length === 0}
+                      placeholder={t('data:create_table.name_placeholder')}
+                    />
+                    <FormErrorOrDescription errors={getFieldErrors(field.state.meta.errors)} />
+                  </div>
+                )}
+              </form.Field>
+              <form.Field
+                name="description"
+                validators={{
+                  onBlur: createTableValueSchema.shape.description,
+                  onChange: createTableValueSchema.shape.description,
+                }}
+              >
+                {(field) => (
+                  <div className="flex flex-col gap-2">
+                    <FormLabel name={field.name}>{t('data:description')}</FormLabel>
+                    <FormInput
+                      type="text"
+                      name={field.name}
+                      defaultValue={field.state.value as string}
+                      onChange={(e) => field.handleChange(e.currentTarget.value)}
+                      onBlur={field.handleBlur}
+                      valid={field.state.meta.errors.length === 0}
+                      placeholder={t('data:create_table.description_placeholder')}
+                    />
+                    <FormErrorOrDescription errors={getFieldErrors(field.state.meta.errors)} />
+                  </div>
+                )}
+              </form.Field>
+            </div>
+            <div className="flex flex-1 flex-row gap-2">
+              <Modal.Close asChild>
+                <Button className="flex-1" variant="secondary">
+                  {t('common:cancel')}
+                </Button>
+              </Modal.Close>
+              <Button className="flex-1" variant="primary" type="submit" name="create">
+                {t('data:create_table.button_accept')}
+              </Button>
+            </div>
+          </div>
+        </form>
+      </Modal.Content>
+    </Modal.Root>
+  );
+}
