+++ packages/app-builder/src/routes/ressources+/data+/createField.tsx
@@ -1,56 +1,11 @@
-import { FormErrorOrDescription } from '@app-builder/components/Form/Tanstack/FormErrorOrDescription';
-import { FormInput } from '@app-builder/components/Form/Tanstack/FormInput';
-import { FormLabel } from '@app-builder/components/Form/Tanstack/FormLabel';
 import { setToastMessage } from '@app-builder/components/MarbleToaster';
-import { EnumDataTypes, isStatusConflictHttpError, UniqueDataTypes } from '@app-builder/models';
+import { isStatusConflictHttpError } from '@app-builder/models';
+import { createFieldValueSchema } from '@app-builder/queries/data/create-field';
 import { initServerServices } from '@app-builder/services/init.server';
 import { captureUnexpectedRemixError } from '@app-builder/services/monitoring';
-import { getFieldErrors } from '@app-builder/utils/form';
 import { getRoute } from '@app-builder/utils/routes';
 import { type ActionFunctionArgs, json } from '@remix-run/node';
-import { useFetcher } from '@remix-run/react';
-import { useForm, useStore } from '@tanstack/react-form';
-import { type Namespace } from 'i18next';
-import { useEffect, useState } from 'react';
-import { useTranslation } from 'react-i18next';
-import { Button, Checkbox, Modal, Select } from 'ui-design-system';
 import { z } from 'zod/v4';
-
-export const handle = {
-  i18n: ['data', 'navigation', 'common'] satisfies Namespace,
-};
-
-const createFieldFormSchema = z.object({
-  name: z
-    .string()
-    .min(1)
-    .regex(/^[a-z]+[a-z0-9_]+$/, {
-      error: 'Only lower case alphanumeric and _, must start with a letter',
-    })
-    .refine((value) => value !== 'id', {
-      error: 'The name "id" is reserved',
-    }),
-  description: z.string(),
-  required: z.string(),
-  type: z.enum(['String', 'Bool', 'Timestamp', 'Float', 'Int']),
-  tableId: z.string(),
-  isEnum: z.boolean(),
-  isUnique: z.boolean(),
-});
-
-type CreateFieldForm = z.infer<typeof createFieldFormSchema>;
-
-const VALUE_TYPES = [
-  { value: 'String', display: 'data:create_field.type_string' },
-  { value: 'Bool', display: 'data:create_field.type_bool' },
-  { value: 'Timestamp', display: 'data:create_field.type_timestamp' },
-  { value: 'Float', display: 'data:create_field.type_float' },
-] as const;
-
-const REQUIRED_OPTIONS = [
-  { value: 'optional', display: 'data:create_field.option_optional' },
-  { value: 'required', display: 'data:create_field.option_required' },
-] as const;
 
 export async function action({ request }: ActionFunctionArgs) {
   const {
@@ -68,7 +23,7 @@ export async function action({ request }: ActionFunctionArgs) {
     }),
   ]);
 
-  const { success, error, data } = createFieldFormSchema.safeParse(raw);
+  const { success, error, data } = createFieldValueSchema.safeParse(raw);
 
   if (!success) return json({ success: 'false', errors: z.treeifyError(error) });
   const { name, description, type, required, tableId, isEnum, isUnique } = data;
@@ -109,252 +64,4 @@ export async function action({ request }: ActionFunctionArgs) {
       );
     }
   }
-}
-
-export function CreateField({ tableId, children }: { tableId: string; children: React.ReactNode }) {
-  const [isOpen, setIsOpen] = useState(false);
-
-  return (
-    <Modal.Root open={isOpen} onOpenChange={setIsOpen}>
-      <Modal.Trigger asChild>{children}</Modal.Trigger>
-      <Modal.Content>
-        <CreateFieldContent
-          tableId={tableId}
-          closeModal={() => {
-            setIsOpen(false);
-          }}
-        />
-      </Modal.Content>
-    </Modal.Root>
-  );
-}
-
-function CreateFieldContent({ tableId, closeModal }: { tableId: string; closeModal: () => void }) {
-  const { t } = useTranslation(handle.i18n);
-  const fetcher = useFetcher<typeof action>();
-
-  const form = useForm({
-    defaultValues: {
-      required: REQUIRED_OPTIONS[0].value,
-      name: '',
-      description: '',
-      type: VALUE_TYPES[0].value,
-      tableId: tableId,
-      isEnum: false,
-      isUnique: false,
-    } as CreateFieldForm,
-    onSubmit: ({ value, formApi }) => {
-      if (formApi.state.isValid) {
-        fetcher.submit(value, {
-          method: 'POST',
-          action: getRoute('/ressources/data/createField'),
-          encType: 'application/json',
-        });
-      }
-    },
-    validators: {
-      onSubmit: createFieldFormSchema,
-    },
-  });
-
-  const selectedType = useStore(form.store, (state) => state.values.type);
-  const selectedEnum = useStore(form.store, (state) => state.values.isEnum);
-  const selectedUnique = useStore(form.store, (state) => state.values.isUnique);
-
-  useEffect(() => {
-    if (fetcher.state === 'idle' && fetcher.data?.success) {
-      closeModal();
-    }
-  }, [closeModal, fetcher.data?.success, fetcher.state]);
-
-  return (
-    <form
-      onSubmit={(e) => {
-        e.preventDefault();
-        e.stopPropagation();
-        form.handleSubmit();
-      }}
-    >
-      <Modal.Title>{t('data:create_field.title')}</Modal.Title>
-      <div className="flex flex-col gap-6 p-6">
-        <div className="flex flex-1 flex-col gap-4">
-          <form.Field
-            name="name"
-            validators={{
-              onChange: createFieldFormSchema.shape.name,
-              onBlur: createFieldFormSchema.shape.name,
-            }}
-          >
-            {(field) => (
-              <div className="flex flex-col gap-2">
-                <FormLabel name={field.name}>{t('data:field_name')}</FormLabel>
-                <FormInput
-                  type="text"
-                  name={field.name}
-                  defaultValue={field.state.value as string}
-                  onChange={(e) => field.handleChange(e.currentTarget.value)}
-                  onBlur={field.handleBlur}
-                  valid={field.state.meta.errors.length === 0}
-                  placeholder={t('data:create_field.name_placeholder')}
-                />
-                <FormErrorOrDescription errors={getFieldErrors(field.state.meta.errors)} />
-              </div>
-            )}
-          </form.Field>
-          <form.Field
-            name="description"
-            validators={{
-              onChange: createFieldFormSchema.shape.description,
-              onBlur: createFieldFormSchema.shape.description,
-            }}
-          >
-            {(field) => (
-              <div className="flex flex-col gap-2">
-                <FormLabel name={field.name}>{t('data:description')}</FormLabel>
-                <FormInput
-                  type="text"
-                  name={field.name}
-                  defaultValue={field.state.value as string}
-                  onChange={(e) => field.handleChange(e.currentTarget.value)}
-                  onBlur={field.handleBlur}
-                  valid={field.state.meta.errors.length === 0}
-                  placeholder={t('data:create_field.description_placeholder')}
-                />
-                <FormErrorOrDescription errors={getFieldErrors(field.state.meta.errors)} />
-              </div>
-            )}
-          </form.Field>
-          <div className="flex flex-row justify-around gap-2">
-            <form.Field
-              name="required"
-              validators={{
-                onChange: createFieldFormSchema.shape.required,
-                onBlur: createFieldFormSchema.shape.required,
-              }}
-            >
-              {(field) => (
-                <div className="flex flex-1 flex-col gap-2">
-                  <FormLabel name={field.name}>{t('data:create_field.required')}</FormLabel>
-                  <Select.Default
-                    className="w-full overflow-hidden"
-                    defaultValue={field.state.value}
-                    onValueChange={(type) => {
-                      field.handleChange(type);
-                    }}
-                  >
-                    {REQUIRED_OPTIONS.map(({ value, display }) => {
-                      return (
-                        <Select.DefaultItem key={value} value={value}>
-                          {t(display)}
-                        </Select.DefaultItem>
-                      );
-                    })}
-                  </Select.Default>
-                  <FormErrorOrDescription errors={getFieldErrors(field.state.meta.errors)} />
-                </div>
-              )}
-            </form.Field>
-            <form.Field
-              name="type"
-              validators={{
-                onChange: createFieldFormSchema.shape.type,
-                onBlur: createFieldFormSchema.shape.type,
-              }}
-            >
-              {(field) => (
-                <div className="flex flex-1 flex-col gap-2">
-                  <FormLabel name={field.name}>{t('data:create_field.type')}</FormLabel>
-                  <Select.Default
-                    className="w-full overflow-hidden"
-                    defaultValue={field.state.value}
-                    onValueChange={(type) => {
-                      field.handleChange(type as (typeof VALUE_TYPES)[number]['value']);
-                    }}
-                  >
-                    {VALUE_TYPES.map(({ value, display }) => {
-                      return (
-                        <Select.DefaultItem key={value} value={value}>
-                          {t(display)}
-                        </Select.DefaultItem>
-                      );
-                    })}
-                  </Select.Default>
-                  <FormErrorOrDescription errors={getFieldErrors(field.state.meta.errors)} />
-                </div>
-              )}
-            </form.Field>
-          </div>
-          {EnumDataTypes.includes(selectedType) ? (
-            <form.Field
-              name="isEnum"
-              validators={{
-                onChange: createFieldFormSchema.shape.isEnum,
-                onBlur: createFieldFormSchema.shape.isEnum,
-              }}
-            >
-              {(field) => (
-                <div className="flex flex-row items-center gap-4">
-                  <Checkbox
-                    name={field.name}
-                    defaultChecked={field.state.value}
-                    disabled={selectedUnique}
-                    onCheckedChange={(checked) => {
-                      if (checked !== 'indeterminate') field.handleChange(checked);
-                    }}
-                  />
-                  <FormLabel name={field.name}>
-                    <p>{t('data:create_field.is_enum.title')}</p>
-                    <p className="text-xs">{t('data:create_field.is_enum.subtitle')}</p>
-                  </FormLabel>
-                  <FormErrorOrDescription errors={getFieldErrors(field.state.meta.errors)} />
-                </div>
-              )}
-            </form.Field>
-          ) : null}
-          {UniqueDataTypes.includes(selectedType) ? (
-            <form.Field
-              name="isUnique"
-              validators={{
-                onChange: createFieldFormSchema.shape.isUnique,
-                onBlur: createFieldFormSchema.shape.isUnique,
-              }}
-            >
-              {(field) => (
-                <div className="flex flex-row items-center gap-4">
-                  <Checkbox
-                    name={field.name}
-                    defaultChecked={field.state.value}
-                    disabled={selectedEnum}
-                    onCheckedChange={(checked) => {
-                      if (checked !== 'indeterminate') field.handleChange(checked);
-                    }}
-                  />
-                  <FormLabel name={field.name}>
-                    <p>{t('data:edit_field.is_unique.title')}</p>
-                    <p className="text-xs">{t('data:edit_field.is_unique.toggle')}</p>
-                    {field.state.value ? (
-                      <p className="text-red-47 text-xs">
-                        {t('data:edit_field.is_unique.warning_creation_asynchronous')}
-                      </p>
-                    ) : null}
-                  </FormLabel>
-                  <FormErrorOrDescription errors={getFieldErrors(field.state.meta.errors)} />
-                </div>
-              )}
-            </form.Field>
-          ) : null}
-        </div>
-        <div className="flex flex-1 flex-row gap-2">
-          <Modal.Close asChild>
-            <Button className="flex-1" variant="secondary">
-              {t('common:cancel')}
-            </Button>
-          </Modal.Close>
-          <Button className="flex-1" variant="primary" type="submit" name="create">
-            {t('data:create_field.button_accept')}
-          </Button>
-        </div>
-      </div>
-    </form>
-  );
 }
