+++ packages/app-builder/src/routes/ressources+/data+/createTable.tsx
@@ -1,35 +1,10 @@
-import { FormErrorOrDescription } from '@app-builder/components/Form/Tanstack/FormErrorOrDescription';
-import { FormInput } from '@app-builder/components/Form/Tanstack/FormInput';
-import { FormLabel } from '@app-builder/components/Form/Tanstack/FormLabel';
 import { setToastMessage } from '@app-builder/components/MarbleToaster';
 import { isStatusConflictHttpError } from '@app-builder/models';
+import { createTableValueSchema } from '@app-builder/queries/data/create-table';
 import { initServerServices } from '@app-builder/services/init.server';
-import { getFieldErrors } from '@app-builder/utils/form';
 import { getRoute } from '@app-builder/utils/routes';
 import { type ActionFunctionArgs, json } from '@remix-run/node';
-import { useFetcher } from '@remix-run/react';
-import { useForm } from '@tanstack/react-form';
-import { type Namespace } from 'i18next';
-import { useEffect, useState } from 'react';
-import { useTranslation } from 'react-i18next';
-import { Button, Modal } from 'ui-design-system';
 import { z } from 'zod/v4';
-
-export const handle = {
-  i18n: ['data', 'navigation', 'common'] satisfies Namespace,
-};
-
-const createTableFormSchema = z.object({
-  name: z
-    .string()
-    .min(1)
-    .regex(/^[a-z]+[a-z0-9_]+$/, {
-      error: 'Only lower case alphanumeric and _, must start with a letter',
-    }),
-  description: z.string(),
-});
-
-type CreateTableForm = z.infer<typeof createTableFormSchema>;
 
 export async function action({ request }: ActionFunctionArgs) {
   const {
@@ -47,7 +22,7 @@ export async function action({ request }: ActionFunctionArgs) {
     }),
   ]);
 
-  const { success, error, data } = createTableFormSchema.safeParse(raw);
+  const { success, error, data } = createTableValueSchema.safeParse(raw);
 
   if (!success) return json({ success: 'false', errors: z.treeifyError(error) });
 
@@ -68,114 +43,4 @@ export async function action({ request }: ActionFunctionArgs) {
       { headers: { 'Set-Cookie': await commitSession(session) } },
     );
   }
-}
-
-export function CreateTable({ children }: { children: React.ReactNode }) {
-  const { t } = useTranslation(handle.i18n);
-  const fetcher = useFetcher<typeof action>();
-
-  const form = useForm({
-    defaultValues: {
-      name: '',
-      description: '',
-    } as CreateTableForm,
-    onSubmit: ({ value, formApi }) => {
-      if (formApi.state.isValid) {
-        fetcher.submit(value, {
-          method: 'POST',
-          action: getRoute('/ressources/data/createTable'),
-          encType: 'application/json',
-        });
-      }
-    },
-    validators: {
-      onSubmit: createTableFormSchema,
-    },
-  });
-
-  const [isOpen, setIsOpen] = useState(false);
-  useEffect(() => {
-    if (fetcher.state === 'idle' && fetcher.data?.success) {
-      setIsOpen(false);
-      form.reset();
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [fetcher.data?.success, fetcher.state]);
-
-  return (
-    <Modal.Root open={isOpen} onOpenChange={setIsOpen}>
-      <Modal.Trigger asChild>{children}</Modal.Trigger>
-      <Modal.Content>
-        <form
-          onSubmit={(e) => {
-            e.preventDefault();
-            e.stopPropagation();
-            form.handleSubmit();
-          }}
-        >
-          <Modal.Title>{t('data:create_table.title')}</Modal.Title>
-          <div className="flex flex-col gap-6 p-6">
-            <div className="flex flex-1 flex-col gap-4">
-              <form.Field
-                name="name"
-                validators={{
-                  onBlur: createTableFormSchema.shape.name,
-                  onChange: createTableFormSchema.shape.name,
-                }}
-              >
-                {(field) => (
-                  <div className="flex flex-col gap-2">
-                    <FormLabel name={field.name}>{t('data:field_name')}</FormLabel>
-                    <FormInput
-                      type="text"
-                      name={field.name}
-                      defaultValue={field.state.value as string}
-                      onChange={(e) => field.handleChange(e.currentTarget.value)}
-                      onBlur={field.handleBlur}
-                      valid={field.state.meta.errors.length === 0}
-                      placeholder={t('data:create_table.name_placeholder')}
-                    />
-                    <FormErrorOrDescription errors={getFieldErrors(field.state.meta.errors)} />
-                  </div>
-                )}
-              </form.Field>
-              <form.Field
-                name="description"
-                validators={{
-                  onBlur: createTableFormSchema.shape.description,
-                  onChange: createTableFormSchema.shape.description,
-                }}
-              >
-                {(field) => (
-                  <div className="flex flex-col gap-2">
-                    <FormLabel name={field.name}>{t('data:description')}</FormLabel>
-                    <FormInput
-                      type="text"
-                      name={field.name}
-                      defaultValue={field.state.value as string}
-                      onChange={(e) => field.handleChange(e.currentTarget.value)}
-                      onBlur={field.handleBlur}
-                      valid={field.state.meta.errors.length === 0}
-                      placeholder={t('data:create_table.description_placeholder')}
-                    />
-                    <FormErrorOrDescription errors={getFieldErrors(field.state.meta.errors)} />
-                  </div>
-                )}
-              </form.Field>
-            </div>
-            <div className="flex flex-1 flex-row gap-2">
-              <Modal.Close asChild>
-                <Button className="flex-1" variant="secondary">
-                  {t('common:cancel')}
-                </Button>
-              </Modal.Close>
-              <Button className="flex-1" variant="primary" type="submit" name="create">
-                {t('data:create_table.button_accept')}
-              </Button>
-            </div>
-          </div>
-        </form>
-      </Modal.Content>
-    </Modal.Root>
-  );
 }
